+++ sample/src/main/java/com/klinker/android/sliding/sample/FabActivity.java
@@ -16,6 +16,7 @@
 
 package com.klinker.android.sliding.sample;
 
+import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Toast;
@@ -50,6 +51,16 @@ public class FabActivity extends SlidingActivity {
                     }
                 }
         );
+
+        Intent intent = getIntent();
+        if (intent.getBooleanExtra(SampleActivity.ARG_USE_EXPANSION, false)) {
+            expandFromPoints(
+                    intent.getIntExtra(SampleActivity.ARG_EXPANSION_LEFT_OFFSET, 0),
+                    intent.getIntExtra(SampleActivity.ARG_EXPANSION_TOP_OFFSET, 0),
+                    intent.getIntExtra(SampleActivity.ARG_EXPANSION_VIEW_WIDTH, 0),
+                    intent.getIntExtra(SampleActivity.ARG_EXPANSION_VIEW_HEIGHT, 0)
+            );
+        }
     }
 
 }