+++ sample/src/main/java/com/klinker/android/sliding/sample/ImageActivity.java
@@ -16,6 +16,7 @@
 
 package com.klinker.android.sliding.sample;
 
+import android.content.Intent;
 import android.os.Bundle;
 import android.os.Handler;
 
@@ -54,6 +55,16 @@ public class ImageActivity extends SlidingActivity {
         //         setImage(R.drawable.profile_picture);
         //     }
         // }, 500);
+
+        Intent intent = getIntent();
+        if (intent.getBooleanExtra(SampleActivity.ARG_USE_EXPANSION, false)) {
+            expandFromPoints(
+                    intent.getIntExtra(SampleActivity.ARG_EXPANSION_LEFT_OFFSET, 0),
+                    intent.getIntExtra(SampleActivity.ARG_EXPANSION_TOP_OFFSET, 0),
+                    intent.getIntExtra(SampleActivity.ARG_EXPANSION_VIEW_WIDTH, 0),
+                    intent.getIntExtra(SampleActivity.ARG_EXPANSION_VIEW_HEIGHT, 0)
+            );
+        }
     }
 
 }