+++ sample/src/main/java/com/klinker/android/sliding/sample/MenuItemActivity.java
@@ -16,6 +16,7 @@
 
 package com.klinker.android.sliding.sample;
 
+import android.content.Intent;
 import android.os.Bundle;
 import android.view.Menu;
 import android.view.MenuItem;
@@ -42,6 +43,16 @@ public class MenuItemActivity extends SlidingActivity {
                 getResources().getColor(R.color.menu_item_activity_primary_dark)
         );
         setContent(R.layout.activity_content);
+
+        Intent intent = getIntent();
+        if (intent.getBooleanExtra(SampleActivity.ARG_USE_EXPANSION, false)) {
+            expandFromPoints(
+                    intent.getIntExtra(SampleActivity.ARG_EXPANSION_LEFT_OFFSET, 0),
+                    intent.getIntExtra(SampleActivity.ARG_EXPANSION_TOP_OFFSET, 0),
+                    intent.getIntExtra(SampleActivity.ARG_EXPANSION_VIEW_WIDTH, 0),
+                    intent.getIntExtra(SampleActivity.ARG_EXPANSION_VIEW_HEIGHT, 0)
+            );
+        }
     }
 
     /**