+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 ## Changelog
 
+### Version 1.1.0:
+- Add an 'Inbox' style expansion animation to the window
+
+### Version 1.0.1:
+- Fix issue with the expansion when there is no header
+
 ### Version 1.0.0:
 - Initial release and feature set